refactor: stop using import * in __init__.py
