+++ src/urtypes/__init__.py
@@ -20,6 +20,3 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from .registry import *
-from .bytes import *
-from .crypto import *
