+++ src/urtypes/cbor/__init__.py
@@ -22,6 +22,3 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from .data import *
-from .decoder import *
-from .encoder import *
