+++ tests/__init__.py
@@ -19,10 +19,3 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-
-from .crypto.test_account import *
-from .crypto.test_bip39 import *
-from .crypto.test_ec_key import *
-from .crypto.test_hd_key import *
-from .crypto.test_output import *
-from .crypto.test_psbt import *
